Applied patch ID: 2799041 "opal 3.6.2 sip 180 ring process error", from zbyslb.
Apparently some systems send DSP in a 180 response, rather than a 183 response. While there is an implication in RFC 3261 sections 20.1.2 and 20.1.5 that if "in-band" call progress is to happen, then 183 should be used, it does not come right out and say anywhere that you CAN'T do SDp in 180.

So, as usual, we have to accommodate other peoples weirdness and it certainly will not hurt anything to accept SDP on a 180 response.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@22773 023b2edf-31b2-4de3-b41e-bca80c47788f