Added ability to get a SIP dialog "context", which is sufficient information to recreate the dialog from a saved string.
Added a function to utilise the above dialog context to send a BYE to the remote clearing the call.

This feature can be useful for servers that had an "unexpected exit" and various clients it was talking to at the time do not implement the RFC4028 session timers, so continue to try and send media forever. They need to be told to cease and desist when the local restarts.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@23132 023b2edf-31b2-4de3-b41e-bca80c47788f