Fixed bug ID: 2831204 "Invalid request-uri in BYE request", reported anonymously.
The problem was nothing to do with the BYE message though this was the symptom observed. The actual issue was the incorrect parsing of quoted URI header values which might contain angle brackets in a double quoted the field parameter, e.g. "2003" <sip:2003@172.19.240.80:5060;transport=udp>;+sip.instance="<urn:uuid:00000000-0000-1000-8000-00085D19E033>"

This confused the parser and ended up with an incorrect Contact field which propagated through to the BYE message.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@23230 023b2edf-31b2-4de3-b41e-bca80c47788f