Fixed numerous issues with fax negotiation. Now OPAL to OPAL works in all 6 modes: T.38, G.711 & T.38 with G.711 fall back for both SIP and H.323.
In process had to:

Fix statistics in spandsp plug in, did not make distinction between final and intermediate values.

Add the ability to control the RFC2833 capability inclusion and options. It was always added, but now must be in the connections media format list. This will be automatic unless the users application explicitly removes it via the media format mask in OpalManager.

Fixed fall back in H.323 if the Mode Request is rejected.

Fixed OpenPhone display of fax statistics when in G.711 mode.


git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@23846 023b2edf-31b2-4de3-b41e-bca80c47788f