Back ports for various issues, mostly related to handling of REGISTER and un-REGISTER:
Comprehensive check of the locking modes for SIP handlers, to try and find a deadlock. Easy access to the activeSIPHandler list enumeration was removed to make sure every usage is picked up by the compiler and inspected by the programmer. This is also useful for future optimisation of the list.

Fixed problem in SIP handlers of trying to initiate a subscription (or publish) when we are unsubscribing or unsubscribed. If get to those states, then handler is doomed and WILL be removed. You can't then start a new transaction as it will end up on a deleted object.

When sending SIP messages (e.g. NOTIFY) and are in the wrong state, wait a little while for the state to change before giving up.

A SIP handler subscribe request must wait for previous subscribe to complete on some message types, e.g. NOTIFY

Fixed sending of NOTIFY waiting for previous one to go out before sending the next.

Change to not use PTimer in handler state activation as it can deadlock when the timer is destroyed, which will happen as it is a local variable. Not sure exactly how this is deadlocking, it seems to be quite a chain of locks, for the moment we just avoid it.

Fixed incorrect register/subscribe refresh timer calculation, thanks zbyslb

Prevent SIP handlers retry timer from executing during destruction.

Do not pre-emptively remove failed SIP register handler if REGISTER fails. Must be garbage collected later or causes crash due to transactions needing the transport contained in the handler instance.

Fixed failure to retry (after a suitable period) REGISTER/SUBSCRIBE if get transport level timeout.


Fix problem with doing DNS lookup when using explicit proxy or presence agent

Fix problem with PUBLISH using wrong expiry after Interval Too Brief error