Added C++ wrapper classes around "C" API.
It may seem odd to have a C++ wrapper around a "C" API which is itself a wrapper around a C++ API, but sometimes a C++ programmer may wish to access the OPAL system via this simplified API instead of the quite complex one in the base OPAL library. It is also easier to use secondary languages such as Java and Ruby using this abstraction than the pure "C" one.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@23894 023b2edf-31b2-4de3-b41e-bca80c47788f