Fixed issue where bad registrars can return multiple identical entries in the Contact field after doing stupid things like removing the transport address or port number from our proposal. Reported by ted szoczei
Also improved the parsing of the Contact field so if had a display name that looked like "Bloggs, Fred" the comma inside a quote would not confuse it.

Finally, again to work around bad registrars, if the registrar does not put a port number in the returned Contact URI, even though we supplied one, we explicitly add it. THis is due to the rules for URI comparison are that no port supplied is different from using the default port. Which in turn is due to a port-less URI meaning a DNS SRV record look up. However, that sort of look up should never occur in a URI from a Contact field, which should always be an explicit location to avoid circular references. So if port absent, we add the default.