Fixed auto-detection of H.264 streams, sometimes work if SPS was small.
Fixed/changed term "frames" to "packets" for statistics (confusing) and added output video frames and frame rate to a final statistics drop to cout.

Added dumping of RTP payloads to a file, this is a simple concatenation of each RTP packets payload, with one exception, for H.264 packets, we add a start code (0,0,0,1) to each payload to convert from packetised H.264 to streamed H.264.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@24103 023b2edf-31b2-4de3-b41e-bca80c47788f