To fix some interop problems with registrars where the ports presence/absence is not handled correctly. There is a distinction between absence and presence with the default value, that is "sip:fred" is not the same as "sip:fred:5060". Some registrars remove the default port if present, others insert it. Sheesh!
Safest thing to do is always assume an explicit port number for REGISTER contact field and never send an empty port. This should be generally OK, though not perfect. It prevents using an SRV record as the Contact, but as the thing doing the registering is nearly always the thing listening on a port, it is the 99.9% solution.

Also removed SetContact that takes a URI as it is up to the caller to decide the form of sanitisation required based on the context of use.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@24197 023b2edf-31b2-4de3-b41e-bca80c47788f