Try to prevent deadlock when have a INVITE with replaces when all calls are within the same OPAL instance.
When doing a consultation transfer, the REFER has a Replaces field in the Refer-To header. When doing so to an internal call, the call-id, to-tag and from-tag are gotten from the SIPConnection. The to-tag and from-tag appears to be around the wrong way. It was changed to fix a bug someone reported. The code in sipcon.cxx:

              << "%3Bto-tag%3D"   << PURL::TranslateString(sip->GetDialog().GetLocalTag(),  PURL::QueryTranslation)
              << "%3Bfrom-tag%3D" << PURL::TranslateString(sip->GetDialog().GetRemoteTag(), PURL::QueryTranslation);

must agree with the other half in sipep.cxx:

      if (context.GetLocalTag() == to && context.GetRemoteTag() == from) {

or it cannot possibly work. I have no idea why it worked when around the other way for the person that reported it. Possibly two bugs cancelling each other out. In any case, the way it is now is the way it must be.