More TLS tweaks.
Needed a keep-alive feature as opensips shuts down the TLS link after a time (65 seconds) regardless of the REGISTER time to live. So we are constantly re-opening and renegotiating the certificates, along with getting a transport error reported to the application.

As it happens the keep-alive can be useful for NAT routers which might time out before the REGISTER time to live too, so the feature was made global to all transports not just TLS.