Fix for bug ID: 3495610 "Incorrect handling of SRV records for PUBLISH", reported by Anonymous
RFC 5509 indicates that the return value from _pres._sip.foo.org is in fact another domain, which should then be processed via the _sip._udp.foo.org SRV record. That is, two look ups are required.

Also if the first fails, we should then try the second with the original domain.