Applied patch for high profile H.264, thanks Alexander Sbitnev.
The above required changes to the way H.323 capabilities are determined. A previous check in to add a "discriminator" bit to plug in media options definitions was backed out and the same technique as is used for SIP was used instead, utilising merge validation and the "EqualMerge" mode for media options.