Merged revision(s) 28564-28584 from opal/branches/v3_12: More logging for DNS queries. ........ Better flexibility on location of wxrc.exe ........ Removed truncation of Windows sound device names in Direct Sound interface.
Change SSL error to be "AccessDenied" as this is usually what it means.
........
Fixed subtle error in DNS resolver code on windows where copying lists did not correctly copy internal strings for the DNS record list. Need to use Windows API function DnsRecordSetCopy and not internal implementation. Amazing that this did not crash more often that it did, which was quite rarely.
........
Fixed loss of media if do multiple hold/retrieve on H.323. Some old OpenH323 hold handling for Music on Hold getting in the way.

Added pause of transmit media (microphone/camera) when putting remote on hold. The stack does not do this as this is where Music on Hold goes, but OpenPhone does not (yet) support that so it needs to shut down the media or the remote non-intuitively can still hear the local user while on hold.

Pausing media streams on hold causes problems, change to do an "internal transfer" to new "Null Audio" device to prevent microphone going out to held endpoint in OpenPhone.

Fix for "pc:*" using default sound channel device names from OpalPCSSEndPoint