Addressed very long standing issue in that RTCP is not processed when have "transmit only" media. WebRTC makes extensive use of the RTCP back channel (NACK, Force I-Frame etc), and DTLS is 100% dependent on receiving packets. So, need to always have a thread reading the socket(s) on an OpalRTPMediaStream, which means that it becomes a push model rather than a pull model.
The above requires there is always a queue between the packet reading thread pushing data and the media patch thread which has the codec (CPU user) in it. For audio this is just the jitter buffer, for others it is a simple queue. Video will eventually need a specialist jitter buffer too, so added a factory for creating jitter buffers based on OpalMediaType.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@32678 023b2edf-31b2-4de3-b41e-bca80c47788f