Fixed setting of remote media formats on first answer SDP.
This is quite important as various attributes of the media formats, e.g. payload types, etc, must be as per remote. Our offer, of necessity, needs to assume the remote does everything we do, so sets the remote media format list to the same as the local one. After answer, we then know the truth, and should use it.